Offre added

+++ Vendeur.java
@@ -2,7 +2,7 @@ public class Vendeur {
 
     public Vendeur(){}
 
-    public Enchere creerEnchere(String nomProduit, int prixInitial, int tempsDepart, int duree){
+    public Enchere creerEnchere(String nomProduit, double prixInitial, int tempsDepart, int duree){
         return new Enchere(this, nomProduit, prixInitial, tempsDepart, duree);
     }
 }
